Removed duplicate entities and moved address and block to the application layer
