--- src/application/types/Block.ts
@@ -1,9 +1,0 @@
-export class Block {
-  height: number;
-  headerHash: string;
-
-  constructor(height: number, headerHash: string) {
-    this.height = height;
-    this.headerHash = headerHash;
-  }
-}
