--- src/application/types/BlockChain.ts
@@ -1,4 +1,0 @@
-export enum BlockChainType {
-  Chia = 'chia',
-  Test = 'test',
-}
